chore: add base for refactoring swaps onto
